New drive and shooter control systems
Added a new TankDrive command that lets the bot move normally at 50% speed.
You can push to 100% forward from there or 100% backward.
The conventional way to do it would be to start at 0% and go to +=100% from
there.

I also made the shooter wheels ramp up to speed when you pull the trigger and
turn them off once the FeederToStartPoint command is finished.