progresss
Working on the enable/disable feature.

The drive train is enabling and disabling perfectly.

The feeder will work, and it disables.
It's just not re-enabling itself.
I think it's because the feeder doesn't have a default command.

The shooter LOOKS like it is being disabled temporarily.
When I hold the shooter button down the wheels move, and if I hit the
disable button while I am doing that the shooter stops even though I am
still holding down the shooter button. This makes me  think that I am
interuppting the TuneShooter command, but that is it. Once the
StopShooter command is over, it is coming back online.

I also haven't fixed the feeder/shooter sync bug.
I should not be able to fire without the shooter button being depressed.