Cleaning code and made autonomous
Autonomous is pretty straightforward.
It's a single command, and it's interfacing with the subsystems
directly.
It should shoot three frisbees and then stop.
Need to check shooter polarity and add a delay timer so we don't try to
shoot before the frisbee falls into the chamber

I am fixing code logic and enforcing naming conditions across the code.
I got interrupted while doing it so I'll go back to it now